+++ HISTORY.rst
@@ -11,3 +11,4 @@ History
     * Added all code to run pyramid application (rudaporto).
     * Added alembic support and first migration (rudaporto).
     * Added docker support (rudaporto).
+    * Added tags ARRAY of strings to mixins.LibraryItemMixin and respective migration (rudaporto).